pyproject.toml
[project]
name = "code-buddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.15",
    "langchain-core>=1.2.30",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.1.0",
    "langgraph>=1.1.6",
    "python-dotenv>=1.2.2",
]
